extract svelte block folding from typescript folding no document highlight yet, VSCode doesn't seems to be merge results from ts and ours
